requirements.txt
google-genai>=1.55.0
openai>=1.40.0
python-dotenv
